Add files via upload
Release Notes
Fixed missing `rm -rf /tmp/choice`
Added a custom icon for the boot picker on some machines

Option to reboot to install media Automatically

No need for EFI Boot (in theory)

Added check to make sure to don’t createinstallmedia to the booted volume (learned this the hard way)

Sorry that there hasn’t been much going on lately, I’m working on a postautomator for this. Hopefully we can get base system patched reliably. If not, it will be an app ran in the OS